+++ migrate-app/src/resources/login/login.controller.ts
@@ -1,30 +1,17 @@
 import { Controller, Post, Body, UseGuards } from '@nestjs/common';
-import jwt from 'jsonwebtoken';
-import { UsersService } from '../users/users.service';
 import { UserCredentials } from './dto/login.dto';
 import { LoginGuard } from './guard/login.guard';
-import config from '../../common/configuration';
+import { LoginService } from './login.service'
 
 @Controller('login')
 export class LoginController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(private readonly loginService: LoginService) { }
 
   @UseGuards(LoginGuard)
   @Post()
   async login(
     @Body() userCredentials: UserCredentials,
   ): Promise<{ token: string }> {
-    const registered = await this.usersService.findByLogin(
-      userCredentials.login,
-    );
-    const token = jwt.sign(
-      {
-        userId: String(registered?.id),
-        login: userCredentials.login,
-      },
-      String(config().secret),
-      { expiresIn: 60 * 60 },
-    );
-    return { token: token };
+    return await this.loginService.login(userCredentials)
   }
 }
