+++ migrate-app/src/resources/login/login.module.ts
@@ -3,10 +3,11 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { LoginController } from './login.controller';
 import { User } from '../users/users.entity';
 import { UsersService } from '../users/users.service';
+import { LoginService } from './login.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
   controllers: [LoginController],
-  providers: [UsersService],
+  providers: [UsersService, LoginService],
 })
-export class LoginModule {}
+export class LoginModule { }
